codewhisper/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "codewhisper"
version = "0.1.0"
description = "Interactive chatbot CLI for exploring mainframe codebases using llm_providers and skills"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "War Rig Team", email = "team@warrig.dev" }
]
keywords = ["chatbot", "mainframe", "cobol", "llm", "code-exploration", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Documentation",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Environment :: Console",
]

dependencies = [
    # LLM provider (replaces LangChain/LangGraph)
    "llm-providers",
    # Code analysis
    "citadel",
    # Data validation and config
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    # CLI
    "typer>=0.12.0",
    "rich>=13.0.0",
    # Skills loading
    "pyyaml>=6.0",
    "python-frontmatter>=1.0.0",
    # Environment
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.12.0",
    "respx>=0.20.0",
    "mypy>=1.8.0",
    "ruff>=0.3.0",
]

[project.scripts]
codewhisper = "codewhisper.cli:cli"

[project.urls]
Homepage = "https://github.com/andrasfe/war_rig"
Repository = "https://github.com/andrasfe/war_rig"
Issues = "https://github.com/andrasfe/war_rig/issues"

[tool.uv.sources]
llm-providers = { path = "../llm-providers", editable = true }
citadel = { path = "../citadel", editable = true }

[tool.hatch.build.targets.wheel]
packages = ["src/codewhisper"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["codewhisper"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
strict_optional = true
